tui/requirements.txt
PyYAML==6.0
textual==0.55.1
textual-dev==1.5.1
rich
pytest==7.3.1
black
flake8
pylint
yapf
